Change GenerateToken into GenerateId and Id
After talking to Kyle, it seemed that this was a better way to go and LISP,
Clojure and XPath have all tread here, so I feel better about doing it.

Rather than generate numeric identifiers, Flabbergast identifiers are
generated. Also, instead of the identifier being generated based on the
invocation of the function, it is generated based on the tuple.

Change-Id: Ied20bc380ec3a2dcaef1772e806799c0885f48a3