Add ludicrously complicated unboxing system to avoid combinatorial explosion
In code involving string concatenation of lookups, there is an explosion of
types. This attempts to control it by keeping certain values boxed. In general,
those going to be placed in frames or converted to strings can remained boxed
since the type will not affect the result type and therefore the result code
path.

Change-Id: I47abe4dafd0e4443157e46e4c686317c9fdf7479